+++ mkchromecast/preferences.py
@@ -46,8 +46,6 @@ try:
 except ImportError:
     import configparser as ConfigParser # This is for Python3
 
-
-
 def ConfigSectionMap(section):
     config = ConfigParser.RawConfigParser()
     configurations = config_manager()    # Class from mkchromecast.config
@@ -185,13 +183,29 @@ if tray == True:
             self.qcnotifications.setCurrentIndex(notindex)
             self.qcnotifications.activated[str].connect(self.onActivatednotify)
 
-            #self.lbl.move(50, 150)
+            """
+            Search at launch
+            """
+            atlaunch = ["enabled", "disabled"]
+            launchindex = atlaunch.index(self.searchatlaunchconf)
+            self.atlaunch = QLabel('Search at launch', self)
+            self.atlaunch.move(20*self.scale_factor, 184*self.scale_factor)
+            self.qcatlaunch = QComboBox(self)
+            self.qcatlaunch.move(180*self.scale_factor, 184*self.scale_factor)
+            self.qcatlaunch.setMinimumContentsLength(7)
+            for item in atlaunch:
+                self.qcatlaunch.addItem(item)
+            self.qcatlaunch.setCurrentIndex(launchindex)
+            self.qcatlaunch.activated[str].connect(self.onActivatedatlaunch)
 
             self.setGeometry(300*self.scale_factor, 300*self.scale_factor, 300*self.scale_factor, 200*self.scale_factor)
-            self.setFixedSize(300*self.scale_factor, 200*self.scale_factor)     #This is to fix the size of the window
+            self.setFixedSize(300*self.scale_factor, 220*self.scale_factor)     #This is to fix the size of the window
             self.setWindowFlags(QtCore.Qt.WindowMinimizeButtonHint | QtCore.Qt.WindowStaysOnTopHint)
             self.setWindowTitle('mkchromecast Preferences')
 
+            """
+            Methods
+            """
         def onActivatedbk(self, text):
             self.config.read(self.configf)
             self.config.set('settings','backend',text)
@@ -261,8 +275,13 @@ if tray == True:
             with open(self.configf, 'w') as configfile:
                     self.config.write(configfile)
             self.read_defaults()
-            #self.lbl.setText(text)
-            #self.lbl.adjustSize()
+
+        def onActivatedatlaunch(self, text):
+            self.config.read(self.configf)
+            self.config.set('settings','searchatlaunch',text)
+            with open(self.configf, 'w') as configfile:
+                    self.config.write(configfile)
+            self.read_defaults()
 
         def read_defaults(self):
             self.backendconf = ConfigSectionMap("settings")['backend']
@@ -276,8 +295,9 @@ if tray == True:
             self.bitrateconf = ConfigSectionMap("settings")['bitrate']
             self.samplerateconf = ConfigSectionMap("settings")['samplerate']
             self.notificationsconf = ConfigSectionMap("settings")['notifications']
+            self.searchatlaunchconf = ConfigSectionMap("settings")['searchatlaunch']
             if debug == True:
-                print(self.backendconf, self.codecconf, self.bitrateconf, self.samplerateconf, self.notificationsconf)
+                print(self.backendconf, self.codecconf, self.bitrateconf, self.samplerateconf, self.notificationsconf, self.searchatlaunchconf)
 
 if __name__ == '__main__':
     app = QApplication(sys.argv)